requirements-serving.txt
# requirements-serving.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0

pydantic>=2.8.0
python-multipart>=0.0.9
requests>=2.32.0

loguru>=0.7.2
prometheus-client>=0.20.0
opentelemetry-api>=1.26.0
opentelemetry-sdk>=1.26.0

numpy>=1.26.0
scikit-learn>=1.5.0
lightgbm>=4.4.0

torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
transformers>=4.46.0
tokenizers>=0.20.0
sentencepiece>=0.2.0
huggingface-hub>=0.25.0
accelerate>=0.33.0

prometheus-fastapi-instrumentator>=6.0.0
bitsandbytes>=0.42.0
accelerate>=0.27.0

opentelemetry-instrumentation-fastapi>=0.47b0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-exporter-otlp-proto-grpc>=1.26.0